A store one side makes is a difference, and umin says it is unsigned
The histogram bin computed as zero got past the equivalence tier: a store or an
atomic is named by where it went, and the comparison looked only at the keys
both sides had. A store one side makes and the other does not is a difference
now.

That found histogram's InterlockedMax coming back as atomic_imax where the
original had atomic_umax, since imin and umin were parsed as one node and the
value declared int. The node says which now, the way ushr does, the variable is
declared uint for it, and the instruction writer casts the value of an unsigned
atomic. particle_update's velocity written back a member at a time has fxc drop
two stores of what was loaded from the same element; the buffer ends up the
same, and that one is a known difference.

+++ HlslDecompiler.Tests/ShaderSources/cs_5_0/tile_luminance.fx
@@ -43,7 +43,7 @@ void main(CS_IN i)
 	}
 	if (i.sv_groupindex == 0) {
 		luminance[i.sv_groupid.xy] = 0.015625 * g1[0];
-		InterlockedAdd(histogram[min(g0[0], 63)], 1);
+		InterlockedAdd(histogram[min((uint)g0[0], 63)], 1);
 	}
 	if ((all(t2 < size) && step == 0) != 0) {
 		luminance[t2] = t3;
